job_portal_major_project/requirements.txt
fastapi[standard]
uvicorn
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlmodel
pydantic
python-dotenv
asyncpg
psycopg2-binary
fastapi-pagination
pytest
pytest-asyncio
httpx